Сделал вывод среднего значения для всех сессий относящихся к одному Kit (отношение ошибок к верным ответам)

+++ app/src/main/java/database/Dao/SessionDao.java
@@ -11,7 +11,7 @@ import database.entities.Session;
 @Dao
 public interface SessionDao {
     @Query("SELECT * FROM session WHERE kit_id == :kitId")
-    List<Session> getAllSessionsByKitId(int kitId);
+    List<Session> getAllSessionsByKitId(long kitId);
     //@Query("SELECT * FROM session WHERE is_succeed == 1")
     //List<Session> getAllSucceedSessions();
     //Все сессии за определенную дату
